main to dev

+++ src/main/(posts)/likes/like.service.ts
@@ -4,28 +4,19 @@ import { LikeRepository } from "./like.repository";
 
 @Injectable()
 export class LikeService {
-  constructor(private readonly likeRepository: LikeRepository) {}
+    constructor(private readonly likeRepository: LikeRepository) {}
 
-  async likePost(userId: string, dto: CreateLikeDto) {
-    // check if already liked or not
-    const exist = await this.likeRepository.alreadyLiked(
-      userId,
-      dto.postId,
-      dto?.commentId,
-    );
-    // if already liked then we have have to dislike that
-    if (exist)
-      return await this.likeRepository.removeLike(
-        userId,
-        dto.postId,
-        dto?.commentId,
-      );
+    async likePost(userId: string, dto: CreateLikeDto) {
+        // check if already liked or not
+        const exist = await this.likeRepository.alreadyLiked(userId, dto.postId, dto?.commentId);
+        // if already liked then we have have to dislike that
+        if (exist) return await this.likeRepository.removeLike(userId, dto.postId, dto?.commentId);
 
-    // if not then create like
-    return await this.likeRepository.like({ ...dto, userId });
-  }
+        // if not then create like
+        return await this.likeRepository.like({ ...dto, userId });
+    }
 
-  async getPostLikes(postId: string) {
-    return await this.likeRepository.getLikesForPost(postId);
-  }
+    async getPostLikes(postId: string) {
+        return await this.likeRepository.getLikesForPost(postId);
+    }
 }
